Fix bidirectional and client streaming listener and avoid duplicates
